Initial support for JSpecify's @NullMarked annotation.
This is the first step towards NullAway adoption of the JSpecify
standard.

Our support for JSpecify must be understood as extremely EXPERIMENTAL.

Currently, all we do is ensure that we treat any `@NullMarked`
third-party packages as `@NonNull`-by-default, using the existing
semantics NullAway applies to "annotated packages" (as passed in
EP flags). We have made no attempt to match the proper semantics
of JSpecify yet, but we are open to looking at issues where NullAway's
behavior differs from the standard.

Note that, in particular, we are aware NullAway doesn't support type
use annotations/annotations in generics. Long term, that's on the roadmap,
but it will likely take a while.

This PR also makes NullAway dependant on the JSpecify annotations
artifact. We could (and might, if needed), remove that dependency
by looking at the annotation's names at runtime as strings, same as
we do for `@Nullable`, but the JSpecify annotations artifact is intended
to be fairly lightweight and we are also expecting it to long-term
become the default way to annotate code for analysis by NullAway.